Refactors number of master and worker nodes to a sub-property of a new Master/Worker type.
Trying to keep commits small I move just the one property first. I chose
the name "replicas" in order to match the vocabulary from k8s.

This change does make obvious one potential down-side to this change and
that is that if the new parent element is missing from the configmap the
parsing library is no longer able to provide the default value, requiring
a duplication of the defaults in code that navigates to the master/worker
configuration. Who do you want a more pleasant experience for, those
using the operator or those writing the operator?

This is a breaking change to the format and so existing configs would
break if this is merged. I assume that is ok as the tool has not yet
reached v1.

Code still needs to be made nullsafe but I think I can wait for discussion
based on a WIP-PR first.