Moves memory and cpu limit configuration to a limits object under the worker and master properties.
The existing code base sets the same values for both the resource request
and resource limit. Was this intentional behaviour? If yes I will remove
the limits object and just place memory and cpu right on the worker and
master nodes. If however it makes sense to allow the resource request
and the resource limit to be configured independently then we can add
a requests object with the same definition and adjust the mapping
accordingly.

Regarding the mapping code I'm not personally a fan of the `isMaster`
boolean parameter, I'd like to refactor this but wanted to keep the scope
of the WIP-PR minimal for discussion of the API changes independently.

Again, code is not finished, needs better tests and some clean up.