Removed unnecessary level of indirection with limits.
As indicated in discussion on the PR it was intentional that the same
values are used for both resource request and limit. This commit
refactors previous work on this to remove the intermediate `Limits`
object.

As part of the process I rewrote some of the null-checking to use a more
functional style.

There is some obvious duplication here but I want to leave it in place
for now to highlight an issue where the name of the generated classes
is not quite as I expected (`worker` field has type `Master`).