Nullsafe all accesses of master and worker data.
In the next version of jsonschema2pojo there is support for mapping
fields that aren't required to `Optional` types. This would provide
a natural mechanism for providing default values when they are missing.
Based on this I've tried to emulate the programming style that this
would provide by wrapping access of optional objects in
`Optional.fromNullable`. When the final version of jsonschema2pojo is
released these "wrappers" can be removed leaving the remaining code
untouched.

Disclaimer, one line with a ternary operator looks particularly ugly,
this will get better with a refactoring I intend to do in a followup PR
 (to not make this one too big).