Fixes conditional running of maven deploy so that it only runs on master builds that are not a pull request.
My PRs are currently failing due to the bash test `[ $TRAVIS_PULL_REQUEST == false ]` returning a non-zero return code, causing travis to mark the build as a failure.

This commit replaces the command chaining syntax with a proper if
statement so that the command still returns a successful code even when
the expression evaluates false.

I also added a check on `$TRAVIS_BRANCH` as I naïvely assume that
snapshots should probably not be published from random branches on the
main repo either (unless the version number has changed in the pom, but
I'm not writing that logic in a bash one-liner).