Fix #187: do not assume the existence of namespace called 'myproject'
Background:
For CRDs we need to create ClusterRoleBinding resource that assings a
ClusterRole to spark-operator ServiceAccount. The service account here
needs to have also the namespace. The only namespace we can assume
to be there is the ns called 'default'. Namespace called 'myproject'
is automatically created in openshift when running oc cluster up.
With K8s or OpenShift installed via Ansible it is not the case.