change SparkApplication default image
This change removes the default image set in the SparkApplication
schema and moves it to the constants definitions. It also adds a
function to detect the environment variable `SPARK_DEFAULT_APP_IMAGE` to
assist in overriding the default. Finally, the README is updated to
include instructions on configuring the operator.