Fix a bug when specifying NoSchedule node tolerations
I discovered a bug that prevents a user from being able to specify Node
Tolerations with the "NoSchedule" effect. These types of tolerations do
not have the "tolerationSeconds" setting, and Kubernetes complains if
you attempt to set them. My previous implementation did not properly
allow for creating a toleration when the tolerationSeconds setting was
not specified.