Include local copies of operator-parent-pom and abstract-operator
This change removes the external dependencies on operator-parent-pom
and abstract-operator by adding copies of them locally to make the
spark-operator easier to maintain and modify. These artifacts will
not be published and will be used only to build the spark-operator
image.

This change also bumps the fabric8 kubernetes client version to 4.10.3.