Include local copies of operator-parent-pom and abstract-operator (#300)
* Include local copies of operator-parent-pom and abstract-operator

This change removes the external dependencies on operator-parent-pom
and abstract-operator by adding copies of them locally to make the
spark-operator easier to maintain and modify. These artifacts will
not be published and will be used only to build the spark-operator
image.

This change also bumps the fabric8 kubernetes client version to 4.10.3.

* Fix path of yaml files for tests after spark-operator source move

* Do not build subprojects as modules

In order to produce an uberjar for spark-operator, the other subprojects
need to be build and install in the local m2 repository as standalone
pieces. Fix the Makefile and pom.xml to work this way.

* Update changelog.md and contributing.md to reflect build changes